+++ libraries/base/changelog.md
@@ -23,6 +23,9 @@
     `ReadS`, as well as related combinators, have been added to
     `Data.Functor.Classes` (#12358)
 
+  * Add `Semigroup` instance for `IO`, as well as for `Event` and `Lifetime`
+    from `GHC.Event` (#12464)
+
 ## 4.9.0.0  *May 2016*
 
   * Bundled with GHC 8.0